Read config file

The thought here is to return a list where the names hew as closely as possible to the names in the standard.

ct_read_config(file)

Arguments

file
character, single line, path to a file or a single string

Value

list, specified in vignette("config")

Examples

  ct_example("keating_1999.CFG") %>%
  ct_read_config()

#> $station_name
#> [1] "4_Victoria_Keating.main_7650"
#> 
#> $rec_dev_id
#> [1] "4_Victoria_Keating.main_7650"
#> 
#> $rev_year
#> [1] 1999
#> Levels: 1991 1999 2013
#> 
#> $TT
#> [1] 8
#> 
#> $`##A`
#> [1] 8
#> 
#> $`##D`
#> [1] 0
#> 
#> $analog_channel
#> # A tibble: 8 × 13
#>      An ch_id    ph                         ccbm     uu           a           b
#>   <int> <chr> <chr>                        <chr>  <chr>       <dbl>       <dbl>
#> 1     1    I1    I1 4_Victoria_Keating.main_7650 ampere -0.01976144   25.373688
#> 2     2    I2    I2 4_Victoria_Keating.main_7650 ampere -0.01977059   25.859930
#> 3     3    I3    I3 4_Victoria_Keating.main_7650 ampere -0.01975111   24.412369
#> 4     4    I4    I4 4_Victoria_Keating.main_7650 ampere -0.27515155 -476.287334
#> 5     5    U1    V1 4_Victoria_Keating.main_7650   volt -0.11381631    8.422407
#> 6     6    U2    V2 4_Victoria_Keating.main_7650   volt -0.11388410    3.074871
#> 7     7    U3    V3 4_Victoria_Keating.main_7650   volt -0.11343370    6.011986
#> 8     8    U4    V4 4_Victoria_Keating.main_7650   volt -0.03804154    0.494540
#> # ... with 6 more variables: skew <dbl>, min <int>, max <int>, primary <dbl>,
#> #   secondary <dbl>, PS <fctr>
#> 
#> $lf
#> [1] 60
#> 
#> $nrates
#> [1] 1
#> 
#> $sampling_rate
#> # A tibble: 1 × 2
#>       samp endsamp
#>      <dbl>   <int>
#> 1 30707.24    2048
#> 
#> $dtm_first
#> [1] "2015-11-18T05:33:03.765333000+00:00"
#> 
#> $dtm_trigger
#> [1] "2015-11-18T05:33:03.790289000+00:00"
#> 
#> $ft
#> [1] ascii
#> Levels: ascii binary
#> 
#> $timemult
#> [1] 1
#>